+++ src/main/java/io/vntr/User.java
@@ -5,13 +5,13 @@ import java.util.Set;
 
 public class User {
     private String name;
-    private Long id;
-    private Set<Long> friendIDs;
+    private Integer id;
+    private Set<Integer> friendIDs;
 
-    public User(String name, Long id) {
+    public User(String name, Integer id) {
         this.name = name;
         this.id = id;
-        this.friendIDs = new HashSet<Long>();
+        this.friendIDs = new HashSet<Integer>();
     }
 
     public String getName() {
@@ -22,23 +22,23 @@ public class User {
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public void befriend(Long friendId) {
+    public void befriend(Integer friendId) {
         friendIDs.add(friendId);
     }
 
-    public void unfriend(Long friendId) {
+    public void unfriend(Integer friendId) {
         friendIDs.remove(friendId);
     }
 
-    public Set<Long> getFriendIDs() {
+    public Set<Integer> getFriendIDs() {
         return friendIDs;
     }
